bioboxmesosscheduler -> de.cebitec package rename
