+++ src/main/java/de/cebitec/mesos/tasks/Task.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.bioboxes.bioboxmesossheduler.tasks;
+package de.cebitec.mesos.tasks;
 
 import java.util.List;
 import org.apache.mesos.Protos;
