+++ src/test/java/de/cebitec/mesos/comparator/DockerTaskComparatorTest.java
